move stripe elements into its own file
